requirements.txt
torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.10.0
opencv-python>=4.5.5.64
dlib>=19.22.0
numpy>=1.19.5
Pillow>=8.0.0
tqdm>=4.50.0
pyyaml>=5.4
transformers
matplotlib
ipykernel
accelerate
scikit-image
pandas
lpips
dreamsim